Change data from long double to double

+++ CS170_Project2_YuvalBar/CS170_Project2_YuvalBar/featureselection.h
@@ -4,7 +4,7 @@
 struct Node
 {
 	int classification;	// 0 or 1
-	std::vector<long double> features;
+	std::vector<double> features;
 };
 
 class FeatureSelection
